backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
